requirements.txt
# Core dependencies for Smart Attack Agent
# Python 3.11+ required

# HTTP clients and web scraping
aiohttp==3.9.3              # Async HTTP client for distributed bots
requests==2.31.0            # Synchronous HTTP for simpler operations
beautifulsoup4==4.12.3      # HTML parsing for crawling
lxml==5.1.0                 # Fast XML/HTML processing
httpx==0.26.0               # Modern HTTP client with HTTP/2 support
supabase==2.4.1             # Supabase client

# URL handling and validation
validators==0.22.0          # URL validation
urllib3==2.2.0             # URL parsing utilities

# Asynchronous support
asyncio==3.4.3             # Async programming

# Data handling
pandas==2.2.0              # CSV processing for credentials
python-dotenv==1.0.1       # Environment variable management

# Reporting
jinja2==3.1.3              # Template engine for HTML reports
markdown==3.5.2            # Markdown to HTML conversion

# Logging and utilities
colorama==0.4.6            # Colored terminal output
tqdm==4.66.1               # Progress bars (sync)
python-dateutil==2.8.2     # Date/time utilities

# Async utilities
asyncio==3.4.3            # Async programming framework
aiofiles==23.2.1          # Async file operations

# Testing (development)
pytest==8.0.0
pytest-asyncio==0.23.5

# Enhanced Network & Browser Support
curl_cffi==0.6.2           # TLS impersonation
playwright==1.41.2         # Headless browser automation
tenacity==8.2.3            # Robust retries
